Convert flag array from int32 to uint8 in v4.
Try to save on some memory and IO transfer costs.

+++ montblanc/impl/biro/v4/BiroSolver.py
@@ -190,7 +190,7 @@ A = [
         test=lambda slvr, ary: rary(ary)),
 
     # Visibility flagging arrays
-    ary_dict('flag', ('ntime', 'nbl', 'nchan', 4), np.int32,
+    ary_dict('flag', ('ntime', 'nbl', 'nchan', 4), np.uint8,
         default=0,
         test=lambda slvr, ary: np.random.random_integers(
             0, 1, size=ary.shape)),
